+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # 0.12.2
 
+Functional changes:
 - Fix on cooldown error messages when spell gets into cooldown
 - Update Spell Binding Table arrow button appearance
+
+API Changes:
+- Add arrow perks damage multiplier
 
 # 0.12.1